--- scripts/generateRss.js
@@ -1,6 +1,0 @@
-/*
- * Copyright (c) Facebook, Inc. and its affiliates.
- */
-const {generateRssFeed} = require('../src/utils/rss');
-
-generateRssFeed();
